+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 * `Toggle` values
+* Multi-use iterators
 
 ## [0.3.0](https://github.com/coady/iterset/releases/tag/v0.3.0) - 2025-01-26
 ### Added